pyproject.toml
[tool.poetry]
name = "django-demo"
version = "0.1.0"
description = ""
authors = ["bradAnderson58 <brad.anderson58@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.2"
django-money = "^2.1.1"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
